mmc: zynq: Use proper size
Error log:
aarch64: +   xilinx_zynqmp
+/home/monstr/data/disk/zynqmp/master/.bm-work/00/drivers/mmc/zynq_sdhci.c:
In function ‘zynq_sdhci_init’:
+/home/monstr/data/disk/zynqmp/master/.bm-work/00/drivers/mmc/zynq_sdhci.c:28:17:
warning: cast to pointer from integer of different size
[-Wint-to-pointer-cast]
+  host->ioaddr = (void *)regbase;

+++ arch/arm/include/asm/arch-zynqmp/sys_proto.h
@@ -13,7 +13,7 @@ static inline void zynq_slcr_gem_clk_setup(u32 gem_id, unsigned long clk_rate)
 {
 }
 
-extern int zynq_sdhci_init(u32 regbase);
+extern int zynq_sdhci_init(unsigned long regbase);
 
 extern unsigned int zynqmp_get_silicon_version(void);
